% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProfiles.R
\name{approxCoordProf}
\alias{approxCoordProf}
\alias{approxProfCoord}
\title{Approximate coordinate profiles with nloptr}
\usage{
approxProfCoord(f, fprime = NULL, optimFun = getAllMaxMin, threshold = 0,
  d, opts = NULL)
}
\arguments{
\item{f}{the function to be evaluated}

\item{fprime}{derivative of the function}

\item{optimFun}{function to do full optimization: at the moment either getAllMaxMin (default choice) or getProfileCoord}

\item{threshold}{threshold of interest}

\item{d}{dimension of the input domain}

\item{opts}{a list containing the options for this function and the subfunctions getMax, getMin or getMaxMinMC, see documentation of getMax, getMin, getMaxMinMC for details. The options only for approxMaxMin are
\itemize{
\item{\code{limits:}}{an optional list with the upper and lower limits of each dimension, if NULL then for each dimension limits are 0,1}
\item{\code{smoother:}}{a string that selects which smoother to use: \itemize{
      \item{\code{"1order"}}: first order interpolation with gradient
      \item{\code{"splineSmooth"}}: smoothing spline with default df (DEFAULT OPTION)
      \item{\code{"quantSpline"}}: quantile spline regression at level 0.95
}}
\item{\code{heavyReturn:}}{If TRUE returns also all minimizers, default is FALSE.}
\item{\code{initDesign:}}{The design of few points where the expensive sup is evaluated.}
\item{\code{fullDesignSize:}}{The full design where the function is approximated.}
\item{\code{preWork:}}{If TRUE, runs a pre-processing procedure to adjust the bad points.}
\item{\code{multistart:}}{number of multistarts for optim procedure.}
\item{\code{MonteCarlo:}}{if TRUE, computes sup with Monte Carlo procedure.}
\item{\code{numMCsamples:}}{number of MC samples for the sup.}
\item{\code{plts:}}{If TRUE, plots the max/min functions at each coordinate, default is FALSE.}
\item{\code{verb:}}{If TRUE, outputs intermediate results, default is FALSE.}
}}
}
\value{
a list of two data frames (min, max) of the evaluations of \eqn{f_sup(x_i) = sup_{x_j \neq i} f(x_1,\dots,x_d) } and \eqn{f_inf(x_i) = inf_{x_j \neq i} f(x_1,\dots,x_d) }
for each i at the design Design. By default Design is a 100 equally spaced points for each dimension. It can be changed by defining it in options$Design
}
\description{
Evaluate profile extrema over other variables with approximations at few values
}
\author{
Dario Azzimonti
}
