% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinateProfiles.R
\name{coordinateProfiles}
\alias{coordinateProfiles}
\title{Coordinate profiles starting from a kriging model}
\usage{
coordinateProfiles(kmModel, threshold, options_full = NULL,
  options_approx = NULL, uq_computations = FALSE, plot_level = 0,
  plot_options = NULL, return_level = 1, ...)
}
\arguments{
\item{kmModel}{a \link[DiceKriging]{km} model}

\item{threshold}{the threshold of interest}

\item{options_full}{an optional list of options for getAllMaxMin, see \link{getAllMaxMin} for details.}

\item{options_approx}{an optional list of options for approxMaxMin, see \link{approxMaxMin} for details.}

\item{uq_computations}{boolean, if TRUE the uq computations for the profile mean are computed.}

\item{plot_level}{an integer to select the plots to return (0=no plots, 1=basic plots, 2= all plots)}

\item{plot_options}{an optional list of parameters for plots. Currently available options
\itemize{
\item{\code{save:}}{boolean, if TRUE saves the plots in \code{folderPlots}}
\item{\code{folderPlots:}}{a string containing the destination folder for plots, if \code{save==TRUE} default is \code{./}}
\item{\code{titleProf:}}{a string containing the title for the coordinate profile plots}
\item{\code{title2d:}}{a string containing the title for the 2d plots (if the input is 2d)}
\item{\code{design:}}{a \eqn{dxr} matrix where \eqn{d} is the input dimension and \eqn{r} is the size of the discretization for plots at each dimension}
\item{\code{id_save:}}{a string to be added to the plot file names, useful for serial computations on HPC.}
\item{\code{qq_fill:}}{if TRUE it fills the region between the first 2 quantiles in \code{quantiles_uq}.}
}}

\item{return_level}{an integer to select the amount of details returned}

\item{...}{additional parameters to be passed to \link{coordProf_UQ}.}
}
\value{
If return_level=1 a list containing
\itemize{
\item{\code{profMean_full:}}{the results of \code{getAllMaxMin} for the posterior mean}
\item{\code{profMean_approx:}}{the results of \code{approxMaxMin} for the posterior mean}
\item{\code{res_UQ:}}{the results of \code{coordProf_UQ} for the posterior mean}
}
if return_level=2 the same list as above but also including
\itemize{
\item{\code{abs_err:}}{the vector of maximum absolute approximation errors for the profile inf /sup on posterior mean for the chosen approximation}
\item{\code{times:} }{ a list containing
	\itemize{
	\item{\code{full:}}{computational time for the full computation of profile extrema}
	\item{\code{approx:}}{computational time for the approximate computation of profile extrema}
	}}
}
}
\description{
The function coordinateProfiles computes the profile extrema functions for the posterior mean of a Gaussian process and its confidence bounds
}
\author{
Dario Azzimonti
}
