% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordProf_UQ.R
\name{coordProf_UQ}
\alias{coordProf_UQ}
\title{Coordinate profiles UQ from a kriging model}
\usage{
coordProf_UQ(object, threshold, allResMean = NULL, quantiles_uq = c(0.05,
  0.95), options_approx = NULL, options_sims = NULL, plot_level = 0,
  plot_options = NULL, return_level = 1)
}
\arguments{
\item{object}{either a \link[DiceKriging]{km} model or a list containing partial results. If \code{object} is a km model then all computations are carried out. If \code{object} is a list, then the function carries out all computations to complete the results list.}

\item{threshold}{the threshold of interest}

\item{allResMean}{a list resulting from \code{getAllMaxMin} or \code{approxMaxMin} for the profile extrema on the mean. If NULL the median from the observations is plotted}

\item{quantiles_uq}{a vector containing the quantiles to be computed}

\item{options_approx}{an optional list of options for approxMaxMin, see \link{approxMaxMin} for details.}

\item{options_sims}{an optional list of options for the posterior simulations.}

\item{plot_level}{an integer to select the plots to return (0=no plots, 1=basic plots, 2= all plots)}

\item{plot_options}{an optional list of parameters for plots. Currently available options
\itemize{
\item{\code{save:}}{boolean, if TRUE saves the plots in \code{folderPlots}}
\item{\code{folderPlots:}}{a string containing the destination folder for plots, if \code{save==TRUE} default is \code{./}}
\item{\code{titleProf:}}{a string containing the title for the coordinate profile plots}
\item{\code{title2d:}}{a string containing the title for the 2d plots (if the input is 2d)}
\item{\code{design:}}{a \eqn{dxr} matrix where \eqn{d} is the input dimension and \eqn{r} is the size of the discretization for plots at each dimension}
\item{\code{id_save:}}{a string to be added to the plot file names, useful for serial computations on HPC.}
\item{\code{qq_fill:}}{if TRUE it fills the region between the first 2 quantiles in \code{quantiles_uq}.}
}}

\item{return_level}{an integer to select the amount of details returned}
}
\value{
If return_level=1 a list containing \itemize{
\item{\code{profSups:}}{an array \code{dxfullDesignSizexnsims} containing the profile sup for each coordinate for each realization.}
\item{\code{profInfs:}}{an array \code{dxfullDesignSizexnsims} containing the profile inf for each coordinate for each realization.}
\item{\code{prof_quantiles_approx:}}{a list containing the quantiles (levels set by \code{quantiles_uq}) of the profile extrema functions.}
} if return_level=2 the same list as above but also including \code{more:} a list containing \itemize{
\item{\code{times:}}{a list containing
	\itemize{
	\item{\code{tSpts:} }{computational time for selecting simulation points.}
	\item{\code{tApprox1ord:}}{vector containing the computational time required for profile extrema computation for each realization}
	}}
\item{\code{simuls:}}{ a matrix containing the value of the field simulated at the simulation points}
\item{\code{sPts:}}{the simulation points}
}
}
\description{
The function coordProf_UQ computes the profile extrema functions for posterior realizations of a Gaussian process and its confidence bounds
}
\author{
Dario Azzimonti
}
