% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMaxNumOpt.R
\name{getAllMaxMin}
\alias{getAllMaxMin}
\title{Coordinate profile extrema with BFGS optimization}
\usage{
getAllMaxMin(f, fprime = NULL, d, options = NULL)
}
\arguments{
\item{f}{the function to be evaluated}

\item{fprime}{derivative of the function}

\item{d}{dimension of the input domain}

\item{options}{a list containing the options for this function and the subfunctions getMax, getMin
see documentation of getMax, getMin for details. The options only for getAllMaxMin are
\itemize{
\item{\code{Design:}}{an optional design matrix with the discretization of each dimension, if NULL then for each dimension Design[,coord] = seq(0,1,length.out=100)}
\item{\code{heavyReturn:}}{If TRUE returns also all minimizers, default is FALSE.}
\item{\code{plts:}}{If TRUE, plots the max/min functions at each coordinate, default is FALSE.}
\item{\code{verb:}}{If TRUE, outputs intermediate results, default is FALSE.}
\item{\code{MonteCarlo:}}{If TRUE, use the MC optimizer otherwise use standard optim.}
}}
}
\value{
a list of two data frames (min, max) of the evaluations of \eqn{f_sup(x_i) = sup_{x_j \neq i} f(x_1,\dots,x_d) } and \eqn{f_inf(x_i) = inf_{x_j \neq i} f(x_1,\dots,x_d) }
for each i at the design Design. By default Design is a 100 equally spaced points for each dimension. It can be changed by defining it in options$Design
}
\description{
Evaluate coordinate profile extrema with full optimization.
}
\author{
Dario Azzimonti
}
