% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMaxNumOpt.R
\name{getClosePoints}
\alias{getClosePoints}
\title{Find close points}
\usage{
getClosePoints(x, allPts, whichDim)
}
\arguments{
\item{x}{one dimensional point}

\item{allPts}{dataframe containing a list of d dimensional points}

\item{whichDim}{integer defining the dimension of x}
}
\value{
the index in allPts (row number) of the closest point in allPts to x along the whichDim dimension
}
\description{
Obtain points close in one specific dimension
}
\author{
Dario Azzimonti
}
