% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProfiles.R
\name{getProfileInf}
\alias{getProfileInf}
\title{Generic profile inf function computation}
\usage{
getProfileInf(eta, Psi, f, fprime, d, options = NULL)
}
\arguments{
\item{eta}{one dimensional point where the function is to be evaluated}

\item{Psi}{projection matrix}

\item{f}{function to be optimized (takes a vector y of dimension d and returns a real number)}

\item{fprime}{derivative of f (same format)}

\item{d}{dimension of the input for f}

\item{options}{a list containing the options to be passed to optim:
\itemize{
\item{\code{par:}}{contains the starting point (a point in dimension d)}
\item{\code{lower:} }{the lower bounds for the optimization domain (see optim for details)}
\item{\code{upper:}}{the upper bounds for the optimization domain (see optim for details)}
}}
}
\value{
a real value corresponding to \eqn{max_{x_1,\dots, x_{coord-1},x_{coord+1}, \dots, x_d} f(x_1,\dots,x_d)}
}
\description{
Compute profile inf function for an arbitrary matrix \code{Psi} with \link[nloptr]{nloptr}.
}
\author{
Dario Azzimonti
}
