% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicsFuns.R
\name{getSegments}
\alias{getSegments}
\title{getSegments}
\usage{
getSegments(y)
}
\arguments{
\item{y}{a vector}
}
\value{
plots the sup and inf of the function for each dimension. If threshold is not NULL
}
\description{
Auxiliary function for \code{getChangePoints}
}
\author{
Dario Azzimonti
}
