% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMaxNumOpt.R
\name{kGradSmooth}
\alias{kGradSmooth}
\title{First order approximation}
\usage{
kGradSmooth(newPoints, profPoints, profEvals, profGradient)
}
\arguments{
\item{newPoints}{vector of points where to approximate the function}

\item{profPoints}{locations where the function was evaluated}

\item{profEvals}{value of the evaluation at profPoints}

\item{profGradient}{value of the gradient at profPoints}
}
\value{
approximated values of the function at newPoints
}
\description{
Compute first order approximation of function from evaluations and gradient
}
\author{
Dario Azzimonti
}
