% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicsFuns.R
\name{plotMaxMin}
\alias{plotMaxMin}
\title{Plot coordinate profiles}
\usage{
plotMaxMin(allRes, Design = NULL, threshold = NULL, trueEvals = NULL,
  km_model = NULL, ...)
}
\arguments{
\item{allRes}{list containing the list \code{res} which contains the computed minima and maxima. The object returned by the function \code{getAllMaxMin}.}

\item{Design}{a d dimensional design corresponding to the points}

\item{threshold}{if not null plots the level}

\item{trueEvals}{if not null adds to each plot the data points and the observed value}

\item{km_model}{if not NULL plots the kriging model trend (TODO uncomment the lines).}

\item{...}{additional parameters to be passed to the plot function}
}
\value{
plots the sup and inf of the function for each dimension. If threshold is not NULL
}
\description{
Plot coordinate profiles, for dimension up to 6.
}
\author{
Dario Azzimonti
}
