% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicsFuns.R
\name{plotOblique}
\alias{plotOblique}
\title{plotOblique}
\usage{
plotOblique(changePoints, direction, ...)
}
\arguments{
\item{changePoints}{Numerical vector with the change points (usually if \code{cp=getChangePoints(...)}, then this is cc$alwaysEx[[1]][[1]] for example)}

\item{direction}{The Psi vector used for the direction}

\item{...}{parameters to be passed to abline}
}
\value{
adds to the current plot the lines \eqn{x} s.t. \code{direction}^T \eqn{x} = \code{changePoints[i]} for all i
}
\description{
Auxiliary function for 2d plotting of excluded regions
}
\author{
Dario Azzimonti
}
