% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profExtrema-package.R
\docType{package}
\name{profExtrema-package}
\alias{profExtrema-package}
\title{profExtrema: Compute and Visualize Profile Extrema Functions}
\description{
Computes profile extrema functions for arbitrary functions. If the function is expensive-to-evaluate it computes profile extrema by emulating the function with a Gaussian process (using package 'DiceKriging'). In this case uncertainty quantification on the profile extrema can also be computed. The different plotting functions for profile extrema give the user a tool to better locate excursion sets.
}
\author{
\strong{Maintainer}: Dario Azzimonti \email{dario.azzimonti@gmail.com} (\href{https://orcid.org/0000-0001-5080-3061}{ORCID}) [copyright holder]

}
\keyword{internal}
