% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{profExtrema}
\alias{profExtrema}
\alias{profExtrema-package}
\title{profExtrema package}
\description{
Computation of profile extrema functions. The package main functions are: \itemize{
    \item \code{\link{coordinateProfiles}}: Given a \link[DiceKriging]{km} objects computes the profile extrema function for the posterior mean an its quantiles.
    \item \code{\link{coordProf_UQ}}: UQ part of \code{coordinateProfiles}.
    \item \code{\link{getAllMaxMin}}: central function to compute the coordinate profile extrema with full optimization.
    \item \code{\link{approxMaxMin}}: central function to compute the coordinate profile extrema with approximations.
}
}
\details{
Package: profExtrema \cr
Type: Package \cr
Version: 0.1.0 \cr
Date: 2017-07-14
}
\references{
TODO

Chevalier, C. (2013). Fast uncertainty reduction strategies relying on Gaussian process models. PhD thesis, University of Bern.
}
\author{
Dario Azzimonti (dario.azzimonti@gmail.com) .
}
