% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_univariate_profiles_UQ.R
\name{setPlotOptions}
\alias{setPlotOptions}
\title{Set-up the plot options when NULL}
\usage{
setPlotOptions(plot_options = NULL, d, num_T)
}
\arguments{
\item{plot_options}{the list of plot options to set-up}

\item{d}{number of coordinates}

\item{num_T}{number of thresholds of interest}
}
\value{
the properly set-up list, if all the fields are already filled then returns \code{plot_options}
}
\description{
function to set-up plot options for plot_univariate_profiles_UQ, coordProf_UQ and coordinateProfiles
}
\author{
Dario Azzimonti
}
