% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProfiles_optim.R
\name{solutionLinConstraint}
\alias{solutionLinConstraint}
\title{Obtain generic solution to linear constraints}
\usage{
solutionLinConstraint(eta, Psi, low = NULL, upp = NULL, pars = NULL)
}
\arguments{
\item{eta}{\eqn{p} dimensional point where the function is to be evaluated}

\item{Psi}{\eqn{p x d} matrix describing the projection matrix}

\item{low}{lower limits of the box constraints on the input (vector of dimension d)}

\item{upp}{upper limits of the box constraints on the input (vector of dimension d)}

\item{pars}{a vector of dimension d-p that can be selected by the user to guide the solution. When NULL (suggested) it is chosen by the function.}
}
\value{
a vector of dimensions \code{d x 1} containing the solution \eqn{x^*}.
}
\description{
Computes an arbitrary solution \eqn{x^*} that satisfies \code{Psi x = eta}
}
\author{
Dario Azzimonti
}
