% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicsFuns.R
\name{plotMaxMin}
\alias{plotMaxMin}
\title{Plot coordinate profiles}
\usage{
plotMaxMin(allRes, Design = NULL, threshold = NULL, changes = FALSE,
  trueEvals = NULL, ...)
}
\arguments{
\item{allRes}{list containing the list \code{res} which contains the computed minima and maxima. The object returned by the function \code{getAllMaxMin}.}

\item{Design}{a d dimensional design corresponding to the points}

\item{threshold}{if not \code{NULL} plots the level}

\item{changes}{boolean, if not \code{FALSE} plots the points where profile extrema take values near the threshold.}

\item{trueEvals}{if not \code{NULL} adds to each plot the data points and the observed value}

\item{...}{additional parameters to be passed to the plot function}
}
\value{
plots the sup and inf of the function for each dimension. If threshold is not NULL
}
\description{
Plot coordinate profiles, for dimension up to 6.
}
\author{
Dario Azzimonti
}
